+++ fxmanifest.lua
@@ -3,7 +3,7 @@ game 'gta5'
 
 author 'Kanayu_u'
 description 'Optimized Doctor NPC for Revive & Heal (Standalone/ox-based)'
-version '1.1.3'
+version '1.2.0'
 
 
 dependency 'ox_lib'
@@ -17,6 +17,7 @@ shared_scripts {
 
 client_scripts {
     '@ox_lib/init.lua',
+    'client/notify.lua',
     'client/utils.lua',
     'client/main.lua'
 }
@@ -28,3 +29,11 @@ server_scripts {
 }
 
 lua54 'yes'
+
+ui_page 'html/notify.html'
+
+files {
+    'html/notify.html',
+    'html/notify.css',
+    'html/notify.js',
+}